+++ src/Tests/FShade.GLSL.Tests/StructsAndClasses.fs
@@ -41,6 +41,11 @@ type CVector2f =
         with get idx = if idx = 0 then this.X else this.Y
         and set idx value = if idx = 0 then this.X <- value else this.Y <- value
 
+type RVector2f =
+    {
+        X : float32
+        Y : float32
+    }
 
 [<ReflectedDefinition>]
 let vectorDot2 (a: Vector2f) (b: Vector2f) =
@@ -68,7 +73,7 @@ let ``Custom struct with constructors``() =
     GLSL.shouldCompileAndContainRegex [ Effect.ofFunction fs ] [
         "Vector2f a;"
         "Vector2f_of_float32\("
-        "Vector2f_of_float32_float32\("
+        "return .+_Vector2f\(v, v\);"
         @"Vector2f_of_Aardvark_Base_V2f\(fs_hugo\.xy\)"
         @"\+ new_.+_Vector2f_of_int32\(3\)\.X"
     ]
@@ -91,7 +96,7 @@ let ``Custom class with constructors``() =
     GLSL.shouldCompileAndContainRegex [ Effect.ofFunction fs ] [
         "CVector2f a = new_.+_CVector2f\(\);"
         "CVector2f_of_float32\("
-        "CVector2f_of_float32_float32\("
+        "return .+_CVector2f\(v, v\)"
         @"CVector2f_of_Aardvark_Base_V2f\(fs_hugo\.xy\)"
         @"\+ new_.+_CVector2f_of_int32\(3\)\.X"
     ]
@@ -132,4 +137,26 @@ let ``Custom class and struct with properties``() =
         @".+_Vector2f one = .+_Vector2f_get_One_.+\(\);"
         @"a\.X = \(\(a\.Y \+ zero\.Y\) \+ one\.X\);"
         @"ColorsOut = \(\(.+_CVector2f_get_Length_.+\(a\) \+ zero\.X\) \+ one\.Y\);"
+    ]
+
+[<Test>]
+let ``Construction with built-in GLSL constructors``() =
+    Setup.Run()
+
+    let fs (v : Vertex) =
+        fragment {
+            let a = { Y = v.hugo.Y; X = v.hugo.X }
+            let b = { a with Y = v.foo.Z }
+            let c, d = v.hugo, v.foo
+            let e = Vector2f(v.hugo.X, v.hugo.Y)
+            let f = CVector2f(v.hugo.X, v.hugo.Y)
+            return a.X + a.Y + b.X + b.Y + c.X + c.Y + d.X + d.Y + e.X + e.Y + f.X + f.Y
+        }
+
+    GLSL.shouldCompileAndContainRegex [ Effect.ofFunction fs ] [
+        ".+_RVector2f a = .+_RVector2f\(fs_hugo\.x, fs_hugo\.y\);"
+        ".+_RVector2f b = .+_RVector2f\(a\.X, fs_foo\.z\);"
+        "= tup_Aardvark_Base_V3f_Aardvark_Base_V4f\(fs_hugo, fs_foo\);"
+        ".+_Vector2f e = .+_Vector2f\(fs_hugo\.x, fs_hugo\.y\);"
+        ".+_CVector2f f = .+_CVector2f\(fs_hugo\.x, fs_hugo\.y\);"
     ]